services/login/requirements.txt
django==5.2.5
djangorestframework==3.16.1
PyJWT==2.10.1
psycopg2-binary==2.9.10
django-cors-headers==4.7.0

